chore: remove header and rename header2
