+++ frontend/src/components/default-layout/index.ts
@@ -1,3 +1,2 @@
 export { default as Footer } from './Footer.vue'
 export { default as Header } from './Header.vue'
-export { default as Header2 } from './Header2.vue'
